Create Playerlist.java
As your repo contains a game based code i would like to add the java code for the player list which maybe helpful for you in game making further for multiplayer gaming. This code doesn't have any library it only contains the logic to fetch the list of the player from the UI and database.